requirements.txt
Django==4.2.1
django-countries==7.5.1
django-phonenumber-field[phonenumbers]==7.1.0
djangorestframework==3.14.0
